+++ src/generate-theme-files.ts
@@ -188,49 +188,51 @@ export const generateThemeFiles = (folder, sectionsSchemas, sectionLocaleCount) 
 
   const target = getAllFiles(folder);
 
-  for (let i = 0; i < target.length; i++) {
-    if (/snippets[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
-      const fileName = target[i].split(/[\\/]/gi).at(-1);
-      const targetFile = snippets.find((sourcePath) =>
-        sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
-      );
-      if (!targetFile) {
-        console.log(
-          `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
-            `Deleted: ${target[i]}`
-          )}`
+  if (process.env.SHOPIFY_CMS_DELETE) {
+    for (let i = 0; i < target.length; i++) {
+      if (/snippets[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
+        const fileName = target[i].split(/[\\/]/gi).at(-1);
+        const targetFile = snippets.find((sourcePath) =>
+          sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
         );
-        fs.unlinkSync(path.join(process.cwd(), target[i]));
+        if (!targetFile) {
+          console.log(
+            `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
+              `Deleted: ${target[i]}`
+            )}`
+          );
+          fs.unlinkSync(path.join(process.cwd(), target[i]));
+        }
       }
-    }
 
-    if (/sections[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
-      const fileName = target[i].split(/[\\/]/gi).at(-1);
-      const targetFile = sections.find((sourcePath) =>
-        sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
-      );
-      if (!targetFile) {
-        console.log(
-          `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
-            `Deleted: ${target[i]}`
-          )}`
+      if (/sections[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
+        const fileName = target[i].split(/[\\/]/gi).at(-1);
+        const targetFile = sections.find((sourcePath) =>
+          sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
         );
-        fs.unlinkSync(path.join(process.cwd(), target[i]));
+        if (!targetFile) {
+          console.log(
+            `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
+              `Deleted: ${target[i]}`
+            )}`
+          );
+          fs.unlinkSync(path.join(process.cwd(), target[i]));
+        }
       }
-    }
 
-    if (/layout[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
-      const fileName = target[i].split(/[\\/]/gi).at(-1);
-      const targetFile = layouts.find((sourcePath) =>
-        sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
-      );
-      if (!targetFile) {
-        console.log(
-          `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
-            `Deleted: ${target[i]}`
-          )}`
+      if (/layout[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
+        const fileName = target[i].split(/[\\/]/gi).at(-1);
+        const targetFile = layouts.find((sourcePath) =>
+          sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
         );
-        fs.unlinkSync(path.join(process.cwd(), target[i]));
+        if (!targetFile) {
+          console.log(
+            `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
+              `Deleted: ${target[i]}`
+            )}`
+          );
+          fs.unlinkSync(path.join(process.cwd(), target[i]));
+        }
       }
     }
   }
