+++ src/index.ts
@@ -388,37 +388,39 @@ export const generateLiquidFiles = (folder: string) => {
     }
   }
 
-  for (let i = 0; i < target.length; i++) {
-    if (
-      /sections[\\/][^\\/]*\.liquid$/gi.test(target[i]) ||
-      /snippets[\\/][^\\/]*\.liquid$/gi.test(target[i])
-    ) {
-      const fileName = target[i].split(/[\\/]/gi).at(-1);
-      const targetFile = source.find((sourcePath) =>
-        sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
-      );
-      if (!targetFile) {
-        console.log(
-          `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
-            `Deleted: ${target[i]}`
-          )}`
+  if (process.env.SHOPIFY_CMS_DELETE) {
+    for (let i = 0; i < target.length; i++) {
+      if (
+        /sections[\\/][^\\/]*\.liquid$/gi.test(target[i]) ||
+        /snippets[\\/][^\\/]*\.liquid$/gi.test(target[i])
+      ) {
+        const fileName = target[i].split(/[\\/]/gi).at(-1);
+        const targetFile = source.find((sourcePath) =>
+          sourcePath.split(/[\\/]/gi).at(-1).includes(fileName)
         );
-        fs.unlinkSync(path.join(process.cwd(), target[i]));
+        if (!targetFile) {
+          console.log(
+            `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
+              `Deleted: ${target[i]}`
+            )}`
+          );
+          fs.unlinkSync(path.join(process.cwd(), target[i]));
+        }
       }
-    }
-    if (/layout[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
-      const fileName = target[i].split(/[\\/]/gi).at(-1);
-      const targetFile = source.find(
-        (sourcePath) =>
-          sourcePath.includes(`layout\\${fileName}`) || sourcePath.includes(`layout/${fileName}`)
-      );
-      if (!targetFile) {
-        console.log(
-          `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
-            `Deleted: ${target[i]}`
-          )}`
+      if (/layout[\\/][^\\/]*\.liquid$/gi.test(target[i])) {
+        const fileName = target[i].split(/[\\/]/gi).at(-1);
+        const targetFile = source.find(
+          (sourcePath) =>
+            sourcePath.includes(`layout\\${fileName}`) || sourcePath.includes(`layout/${fileName}`)
         );
-        fs.unlinkSync(path.join(process.cwd(), target[i]));
+        if (!targetFile) {
+          console.log(
+            `[${chalk.gray(new Date().toLocaleTimeString())}]: ${chalk.redBright(
+              `Deleted: ${target[i]}`
+            )}`
+          );
+          fs.unlinkSync(path.join(process.cwd(), target[i]));
+        }
       }
     }
   }
